requirements.txt
torch>=2.0.0
torchvision>=0.15.0
clip-by-openai
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0
PyQt5>=5.15.0